feat: roll driver, switch to new api.json format (#199)

+++ playwright/src/main/java/com/microsoft/playwright/Mouse.java
@@ -21,7 +21,9 @@ import java.util.*;
 /**
  * The Mouse class operates in main-frame CSS pixels relative to the top-left corner of the viewport.
  * <p>
- * Every {@code page} object has its own Mouse, accessible with page.mouse.
+ * Every {@code page} object has its own Mouse, accessible with [{@code property: Page.mouse}].
+ * <p>
+ * 
  * <p>
  */
 public interface Mouse {
@@ -33,13 +35,13 @@ public interface Mouse {
      */
     public Button button;
     /**
-     * defaults to 1. See UIEvent.detail.
+     * defaults to 1. See [UIEvent.detail].
      */
     public Integer clickCount;
     /**
      * Time to wait between {@code mousedown} and {@code mouseup} in milliseconds. Defaults to 0.
      */
-    public Integer delay;
+    public Double delay;
 
     public ClickOptions withButton(Button button) {
       this.button = button;
@@ -49,7 +51,7 @@ public interface Mouse {
       this.clickCount = clickCount;
       return this;
     }
-    public ClickOptions withDelay(Integer delay) {
+    public ClickOptions withDelay(Double delay) {
       this.delay = delay;
       return this;
     }
@@ -62,13 +64,13 @@ public interface Mouse {
     /**
      * Time to wait between {@code mousedown} and {@code mouseup} in milliseconds. Defaults to 0.
      */
-    public Integer delay;
+    public Double delay;
 
     public DblclickOptions withButton(Button button) {
       this.button = button;
       return this;
     }
-    public DblclickOptions withDelay(Integer delay) {
+    public DblclickOptions withDelay(Double delay) {
       this.delay = delay;
       return this;
     }
@@ -79,7 +81,7 @@ public interface Mouse {
      */
     public Button button;
     /**
-     * defaults to 1. See UIEvent.detail.
+     * defaults to 1. See [UIEvent.detail].
      */
     public Integer clickCount;
 
@@ -109,7 +111,7 @@ public interface Mouse {
      */
     public Button button;
     /**
-     * defaults to 1. See UIEvent.detail.
+     * defaults to 1. See [UIEvent.detail].
      */
     public Integer clickCount;
 
@@ -126,14 +128,16 @@ public interface Mouse {
     click(x, y, null);
   }
   /**
-   * Shortcut for {@code mouse.move(x, y[, options])}, {@code mouse.down([options])}, {@code mouse.up([options])}.
+   * Shortcut for [{@code method: Mouse.move}], [{@code method: Mouse.down}], [{@code method: Mouse.up}].
    */
   void click(int x, int y, ClickOptions options);
   default void dblclick(int x, int y) {
     dblclick(x, y, null);
   }
   /**
-   * Shortcut for {@code mouse.move(x, y[, options])}, {@code mouse.down([options])}, {@code mouse.up([options])}, {@code mouse.down([options])} and {@code mouse.up([options])}.
+   * Shortcut for [{@code method: Mouse.move}], [{@code method: Mouse.down}], [{@code method: Mouse.up}], [{@code method: Mouse.down}] and
+   * <p>
+   * [{@code method: Mouse.up}].
    */
   void dblclick(int x, int y, DblclickOptions options);
   default void down() {
